Refactor `fetch.py` and `main.py`.
This commit simplifies the logic in `get_papers()`, `parse_response()`,
and `main()`. We use lazy iteration instead of loops to cut down on
boilerplate and to make things slightly faster.

The publication date is used everywhere instead of a mix of the last
updated date and publication date, under the assumption that papers
don't change enough to warrant re-rating after initial publication.

We move the query construction into the top-level to separate it from
the downloading logic, and we move the start date determination into
`main()` from `get_papers()`. The `init` date is now defaulted to 3 days
ago in the absence of a checkpoint/save.

* Fix missing stuff and typo in `requirements.txt`.
* Rewrite `Paper.__repr__()` and use it in `interface.py`.
* Drop `batch_size` argument as it is unnecessary for the moment.
* Add an immutable `Save` dataclass to reduce boilerplate in the
  checkpointing procedure.